moved boolean component into tc-ui
