+++ demo/tc-ui/templates/components/edit-components/index.jsx
@@ -1,5 +1,4 @@
 const { primitives } = require('@financial-times/tc-ui/server');
-const Boolean = require('./edit-boolean');
 const DateTime = require('./edit-date-time');
 const Dropdown = require('./edit-dropdown');
 const Text = require('./edit-text');
@@ -7,7 +6,6 @@ const Number = require('./edit-number');
 const TextArea = require('./edit-text-area');
 
 module.exports = {
-	Boolean,
 	DateTime,
 	Dropdown,
 	Text,
