+++ packages/tc-ui/primitives/boolean/server.jsx
@@ -1,0 +1,66 @@
+const { h } = require('preact');
+const { FieldTitle } = require('../../lib/edit-helpers');
+
+const getBooleanLabel = value => {
+	if (value === true) return 'Yes';
+	if (value === false) return 'No';
+	return 'Unknown';
+};
+
+const Checkbox = ({ name, checkboxValue, disabled, userValue }) => {
+	const label = getBooleanLabel(checkboxValue);
+	return (
+		<label>
+			<input
+				type="radio"
+				name={name}
+				value={checkboxValue.toString()}
+				aria-label={label}
+				id={`radio1-${name}`}
+				checked={userValue === checkboxValue ? 'true' : null}
+				disabled={disabled}
+			/>
+
+			<span className="o-forms-input__label" aria-hidden="true">
+				{label}
+			</span>
+		</label>
+	);
+};
+
+const EditBoolean = props => {
+	const { propertyName, value, lockedBy } = props;
+	const disabled = !!lockedBy;
+	return (
+		<div
+			className="o-forms-field"
+			data-biz-ops-type="boolean"
+			role="group"
+			aria-labelledby="inline-radio-round-group-title"
+		>
+			<FieldTitle {...props} />
+			<span className="o-forms-input o-forms-input--radio-round o-forms-input--inline">
+				<Checkbox
+					name={propertyName}
+					checkboxValue
+					disabled={disabled}
+					userValue={value}
+				/>
+				<Checkbox
+					name={propertyName}
+					checkboxValue={false}
+					disabled={disabled}
+					userValue={value}
+				/>
+			</span>
+		</div>
+	);
+};
+
+module.exports = {
+	ViewComponent: ({ value, id }) => (
+		<span id={id}>{getBooleanLabel(value)}</span>
+	),
+	EditComponent: EditBoolean,
+	parser: value => (value === undefined ? undefined : value === 'true'),
+};
