+++ packages/tc-ui/primitives/server.js
@@ -1,5 +1,7 @@
+const Boolean = require('./boolean/server');
 const Relationship = require('./relationship/server');
 
 module.exports = {
+	Boolean,
 	Relationship,
 };
